feat: Add CreateChannelModal component
This commit adds the CreateChannelModal component, which allows users to create a new channel. The component includes a form with input fields for the channel name and type, as well as a "Create" button. When the form is submitted, the values are sent to the server API to create the channel. This enhancement improves the user experience by providing a convenient way to create channels within the application.