feat: Update CreateChannelModal to set default channel type based on data prop
This commit updates the CreateChannelModal component to set the default channel type based on the data prop passed from the useModal hook. If the data prop contains a channelType value, it will be used as the default channel type. Otherwise, the default channel type will be set to ChannelType.TEXT. This enhancement improves the user experience by preselecting the appropriate channel type when creating a new channel.