feat: Add Channel interface to use-modal-store.ts
This commit adds the Channel interface to the use-modal-store.ts file. The Channel interface is imported from the @prisma/client package and is used to define the type of the channel property in the ModalData interface. This enhancement improves the accuracy and type safety of the code when working with channels in the modal store.