feat: Add onClick event to Trash icon in ServerChannel component
This commit adds an onClick event to the Trash icon in the ServerChannel component. When clicked, it triggers the onOpen function from the useModal hook, passing in the "deleteChannel" modal type and the server and channel objects. This enhancement improves the functionality of the ServerChannel component by allowing users to open the delete channel modal when clicking on the Trash icon.